refactor(icons): install icon libraries through the web-asset API
The icon library feature carried four endpoints of its own
(geticonlibraries, addiconlibrary, refreshiconlibrary, deleteiconlibrary)
and an IconLibraries table, right next to the generic web-asset API that
already stores exactly these files. All four are gone; installing, listing
and removing a library now goes through uploadwebasset, getwebassets and
deletewebasset.

- uploadwebasset takes either data= (base64, unchanged) or url=, in which
  case the server downloads it itself. For a stylesheet it follows every
  url() reference as well, stores those files and rewrites the references
  to the stored copies. The downloader is the old IconLibraries.cpp,
  renamed to WebAssetFetch.cpp now that it is not icon-specific, with its
  limits intact: http/https only, a 20 second timeout per request, at most
  32 files and 16 MiB per bundle, 8 MiB per file, and every fetched URL
  logged.
- getwebassets is readable by any logged in user rather than by
  administrators only: a viewer's browser has to know which stylesheets to
  load or every icon taken from one renders as an empty box. It reports
  name, path, LastUpdate and, for administrators only, SourceURL, since
  that can name a host on the local network.
- deletewebasset removes the companion files the asset brought in and
  clears its metadata row, so removing a stylesheet takes its fonts with
  it instead of leaving them behind.
- The IconLibraries table is replaced by a generic WebAssets table (Name,
  SourceURL, Companions, LastUpdate). DB_VERSION 183 is unreleased, so its
  migration was edited in place: it now adds DeviceStatus.Icon and creates
  WebAssets, in both the upgrade and the fresh-install path. Refreshing is
  uploadwebasset with the recorded URL, so no refresh endpoint is left.
- Front-end: iconLibraries.js links every .css asset, the icon picker
  reads its library list from getwebassets and derives the class prefix
  from the filename the library was installed under, and the Custom Icons
  page drives the three endpoints. The ZIP custom icon upload and the icon
  editing are untouched.
