+++ msbuild/domoticz.vcxproj
@@ -761,7 +761,6 @@
     <ClInclude Include="..\main\EventSystem.h" />
     <ClInclude Include="..\main\GZipHelper.h" />
     <ClInclude Include="..\main\HTMLSanitizer.h" />
-    <ClInclude Include="..\main\IconLibraries.h" />
     <ClInclude Include="..\main\IFTTT.h" />
     <ClInclude Include="..\main\json_helper.h" />
     <ClInclude Include="..\main\localtime_r.h" />
@@ -829,6 +828,7 @@
     <ClInclude Include="..\main\stdafx.h" />
     <ClInclude Include="..\main\SunRiseSet.h" />
     <ClInclude Include="..\main\ThemeSettings.h" />
+    <ClInclude Include="..\main\WebAssetFetch.h" />
     <ClInclude Include="..\main\WebAssets.h" />
     <ClInclude Include="..\tcpserver\TCPClient.h" />
     <ClInclude Include="..\tcpserver\TCPServer.h" />
@@ -1063,7 +1063,6 @@
     <ClCompile Include="..\main\EventsPythonModule.cpp" />
     <ClCompile Include="..\main\EventSystem.cpp" />
     <ClCompile Include="..\main\HTMLSanitizer.cpp" />
-    <ClCompile Include="..\main\IconLibraries.cpp" />
     <ClCompile Include="..\main\IFTTT.cpp" />
     <ClCompile Include="..\main\json_helper.cpp" />
     <ClCompile Include="..\main\localtime_r.cpp" />
@@ -1126,6 +1125,7 @@
     <ClCompile Include="..\tcpserver\TCPClient.cpp" />
     <ClCompile Include="..\tcpserver\TCPServer.cpp" />
     <ClCompile Include="..\httpclient\UrlEncode.cpp" />
+    <ClCompile Include="..\main\WebAssetFetch.cpp" />
     <ClCompile Include="..\main\WebServer.cpp" />
     <ClCompile Include="..\main\WebServerCmds.cpp" />
     <ClCompile Include="..\main\WebServerCommands.cpp" />
