+++ www/views/customicons.html
@@ -71,8 +71,8 @@
                 <th></th>
             </tr>
             <tr data-ng-repeat="library in iconlibraries">
-                <td>{{library.Name}}</td>
-                <td>{{library.Prefix}}</td>
+                <td>{{library.name}}</td>
+                <td>{{library.prefix}}</td>
                 <td style="word-break: break-all;">{{library.SourceURL}}</td>
                 <td>{{library.LastUpdate}}</td>
                 <td align="right" style="white-space: nowrap;">
@@ -90,11 +90,7 @@
     <div class="page-content-container">
         <table id="iconlibraryaddtable" border="0" cellpadding="0" cellspacing="10" width="100%">
             <tr>
-                <td align="right" style="width:110px"><label for="iconlibname"><span data-i18n="Name"></span>:</label></td>
-                <td><input type="text" id="iconlibname" ng-model="newLibrary.name" maxlength="100" style="width: 260px; padding: .2em;" class="text ui-widget-content ui-corner-all"></td>
-            </tr>
-            <tr>
-                <td align="right"><label for="iconliburl"><span data-i18n="URL"></span>:</label></td>
+                <td align="right" style="width:110px"><label for="iconliburl"><span data-i18n="URL"></span>:</label></td>
                 <td><input type="text" id="iconliburl" ng-model="newLibrary.url" maxlength="500" style="width: 440px; padding: .2em;" class="text ui-widget-content ui-corner-all"></td>
             </tr>
             <tr>
